+++ Main/SubModule.cs
@@ -127,6 +127,7 @@ public class SubModule : MBSubModuleBase
         var peaceHook = IoC.Resolve<IOnPeaceAction>();
         DiplomacyIoC.InitializeHooks(allianceHook, peaceHook);
         AllianceCampaignBehavior_EndAlliance_Patch.Initialize(logger);
+        AllianceCampaignBehavior_AddAllianceDecision_Patch.Initialize(logger);
         DeclareWarAction_ApplyInternal_Patch.Initialize(logger);
         MakePeaceAction_ApplyInternal_Patch.Initialize(logger);